requirements.txt
torch>=2.0
torchvision>=0.15
click>=8.1.3
matplotlib>=3.7.2
scikit-learn
hydra-core